tests/wpt/web-platform-tests/webmessaging/with-ports/001.html

<!doctype html>
<title>resolving broken url</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_throws_dom('SYNTAX_ERR', function() {
    postMessage('', 'http://foo bar', []);
  }, 'should have failed to resolve');
});
</script>
